Create QUIZ_C_PROGRAMMING
The code is a quiz about informatic and c language.

The available replies're 15 and they're stored in char questions: char question is a string array of three size:[15][6][200] (15 is the number of questions, 6 because question+4 answer options+correct option, 200 is the number max. characters of string).
I'll explain better in the READme about how the question is casually generated and the end of program.